requirements.txt
absl-py==0.12.0
attrs==20.3.0
cachelib==0.1.1
click==7.1.2
dataclasses==0.6
Flask==1.1.2
Flask-Login==0.5.0
Flask-Session==0.3.2
Flask-WTF==0.14.3
flatbuffers==1.12
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpickle==2.0.0
MarkupSafe==1.1.1
mediapipe==0.8.3.1
#MNN==1.1.3
mysql-connector-python==8.0.23
numpy==1.20.1
onnxruntime==1.7.0
opencv-python==4.5.1.48
phonenumbers==8.12.19
Pillow==8.1.2
protobuf==3.15.6
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
